Create PoC II Nuts-LSP
This document describes the technical standards and the use of Verifiable Credentials (VCs) and FHIR for exchanging patient and healthcare provider data between the Nuts network and the LSP (Landelijk Schakelpunt).
To ensure secure, interoperable, and policy-compliant communication, all interaction between the two ecosystems is routed through the GtK (Gevalideerd Twiin Knooppunt).